+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop the unneeded ConfigurationController (#101)
 
 ### Fixed
+- Adapt the usage of core-provided labels to TYPO3 8.7 (#115)
 - Update the BE module icon definition for TYPO3 8.7 (#114)
 - Migrate the TCA wizards for TYPO3 8.7 (#113)
 - Hide the test tables from BE user table permission lists (#112)